+++ layout/style/test/test_transitions_per_property.html
@@ -852,7 +852,7 @@ const basicShapesTests = [
   { start: "circle(500px)", end: "circle(farthest-side)",
     expected: ["circle", ["farthest-side at 50% 50%"]]},
   { start: "circle(500px)", end: "circle(closest-side)",
-    expected: ["circle", ["closest-side at 50% 50%"]]},
+    expected: ["circle", ["at 50% 50%"]]},
   { start: "ellipse()", end: "ellipse(50px 50px)",
     expected: ["ellipse", ["50px 50px at 50% 50%"]] },
   { start: "ellipse(closest-side closest-side)", end: "ellipse(500px 500px)",
