--- testing/web-platform/meta/css/css-shapes/parsing/shape-outside-computed.html.ini
@@ -1,4 +1,0 @@
-[shape-outside-computed.html]
-  [Property shape-outside value 'circle(at 10% 20%)' computes to 'circle(at 10% 20%)']
-    expected: FAIL
-
