+++ testing/web-platform/meta/css/css-shapes/shape-outside/values/shape-outside-circle-000.html.ini
@@ -1,10 +1,4 @@
 [shape-outside-circle-000.html]
-  [No arguments - computed]
-    expected: FAIL
-
-  [Position argument only - computed]
-    expected: FAIL
-
   [No arguments - inline]
     expected: FAIL
