+++ testing/web-platform/meta/css/css-shapes/shape-outside/values/shape-outside-circle-001.html.ini
@@ -1,10 +1,4 @@
 [shape-outside-circle-001.html]
-  [circle() - computed]
-    expected: FAIL
-
-  [circle(closest-side) - computed]
-    expected: FAIL
-
   [circle() - inline]
     expected: FAIL
