+++ testing/web-platform/meta/css/css-shapes/shape-outside/values/shape-outside-circle-009.html.ini
@@ -1,61 +1,4 @@
 [shape-outside-circle-009.html]
-  [circle(at +50px) - computed]
-    expected: FAIL
-
-  [circle(at -50px) - computed]
-    expected: FAIL
-
-  [circle(at +50%) - computed]
-    expected: FAIL
-
-  [circle(at -50%) - computed]
-    expected: FAIL
-
-  [circle(at left +50px) - computed]
-    expected: FAIL
-
-  [circle(at left +50%) - computed]
-    expected: FAIL
-
-  [circle(at right -50px) - computed]
-    expected: FAIL
-
-  [circle(at right -50%) - computed]
-    expected: FAIL
-
-  [circle(at +50px top) - computed]
-    expected: FAIL
-
-  [circle(at +50% top) - computed]
-    expected: FAIL
-
-  [circle(at -50px bottom) - computed]
-    expected: FAIL
-
-  [circle(at -50% bottom) - computed]
-    expected: FAIL
-
-  [circle(at +50px +50px) - computed]
-    expected: FAIL
-
-  [circle(at +50% +50%) - computed]
-    expected: FAIL
-
-  [circle(at -50px -50px) - computed]
-    expected: FAIL
-
-  [circle(at +50px -50px) - computed]
-    expected: FAIL
-
-  [circle(at -50px +50px) - computed]
-    expected: FAIL
-
-  [circle(at +50% -50%) - computed]
-    expected: FAIL
-
-  [circle(at -50% +50%) - computed]
-    expected: FAIL
-
   [circle(at +50px) - inline]
     expected: FAIL
