+++ testing/web-platform/meta/css/css-shapes/shape-outside/values/shape-outside-circle-011.html.ini
@@ -2,27 +2,6 @@
   [circle(at calc((12.5%*6 + 10in) / 4)) - inline style]
     expected: FAIL
 
-  [circle(at calc(10in)) - computed style]
-    expected: FAIL
-
-  [circle(at calc(10in + 20px)) - computed style]
-    expected: FAIL
-
-  [circle(at calc(30%)) - computed style]
-    expected: FAIL
-
-  [circle(at calc(100%/4)) - computed style]
-    expected: FAIL
-
-  [circle(at calc(25%*3)) - computed style]
-    expected: FAIL
-
-  [circle(at calc(25%*3 - 10in)) - computed style]
-    expected: FAIL
-
-  [circle(at calc((12.5%*6 + 10in) / 4)) - computed style]
-    expected: FAIL
-
   [circle(at calc(10in)) - inline style]
     expected: FAIL
